requirements.txt
mem0ai>=0.1.40
transformers>=4.40.0
torch>=2.1.0
accelerate>=0.27.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
huggingface-hub>=0.23.0